third_party/WebKit/LayoutTests/bluetooth/disconnect.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/bluetooth-helpers.js"></script>
<script>
'use strict';

test(t => { assert_true(window.testRunner instanceof Object); t.done(); },
     'window.testRunner is required for the following tests.');

// TODO(ortuno): Write tests to check that "Disconnect" was actually
// called on the device.
// http://crbug.com/569716
promise_test(() => {
  testRunner.setBluetoothMockDataSet('HeartRateAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['heart_rate']}]})
    .then(device => device.gatt.connect())
    .then(gattServer => {
      gattServer.disconnect();
      assert_false(gattServer.connected);
    });
}, '\'connected\' is set to false after disconnect is called.');

promise_test(() => {
  testRunner.setBluetoothMockDataSet('HeartRateAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['heart_rate']}]})
    .then(device => device.gatt.connect())
    .then(gattServer => {
      gattServer.disconnect();
      assert_false(gattServer.connected);
      gattServer.disconnect();
      assert_false(gattServer.connected);
    });
}, 'Calling disconnect twice in a row still results in \'connected\' ' +
   'being false.');

promise_test(() => {
  testRunner.setBluetoothMockDataSet('HeartRateAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['heart_rate']}]})
    .then(device => {
      return device.gatt.connect().then(gattServer => {
        gattServer.disconnect();
        assert_false(gattServer.connected);
      })
      .then(() => device.gatt.connect()).then(gattServer => {
        gattServer.disconnect();
        assert_false(gattServer.connected);
      });
    });
}, 'Connect + Disconnect twice still results in \'connected\' being false.');
</script>
